requirements.txt
pyxel==2.3.0
pymunk-6.11.0
micropip-0.8.0
asyncio-3.4.3
jinja2-3.1.5
beautifulsoup4-4.12.3
